requirements.txt
streamlit
langchain
langgraph
langchain-openai
python-dotenv
apify-client
lxml
beautifulsoup4
langchain_community
fastapi
uvicorn
requests
langgraph-checkpoint-sqlite
lxml-html-clean
pydantic>=2.0.0,<3.0.0
tenacity
